[REFACTOR/#189] core-ui 패키징 변경
