+++ core-ui/src/main/java/com/going/ui/state/EnumUiState.kt
@@ -1,4 +1,4 @@
-package com.going.ui.extension
+package com.going.ui.state
 
 enum class EnumUiState {
     LOADING, SUCCESS, FAILURE, EMPTY
