+++ core-ui/src/main/java/com/going/ui/util/ItemDiffCallback.kt
@@ -1,4 +1,4 @@
-package com.going.ui.extension
+package com.going.ui.util
 
 import androidx.recyclerview.widget.DiffUtil
