+++ presentation/src/main/java/com/going/presentation/dashboard/triplist/completed/CompletedAdapter.kt
@@ -5,7 +5,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
 import com.going.domain.entity.response.DashBoardModel.DashBoardTripModel
 import com.going.presentation.databinding.ItemDashBoardCompletedBinding
-import com.going.ui.extension.ItemDiffCallback
+import com.going.ui.util.ItemDiffCallback
 
 class CompletedAdapter(
     private val listener: OnDashBoardSelectedListener
