+++ presentation/src/main/java/com/going/presentation/entertrip/invitetrip/invitecode/EnterTripActivity.kt
@@ -15,7 +15,7 @@ import com.going.presentation.entertrip.invitetrip.finish.InviteFinishActivity
 import com.going.presentation.entertrip.invitetrip.invitecode.EnterTripViewModel.Companion.ERROR_NO_EXIST
 import com.going.presentation.entertrip.invitetrip.invitecode.EnterTripViewModel.Companion.ERROR_OVER_SIX
 import com.going.ui.base.BaseActivity
-import com.going.ui.extension.UiState
+import com.going.ui.state.UiState
 import com.going.ui.extension.hideKeyboard
 import com.going.ui.extension.setOnSingleClickListener
 import com.going.ui.extension.toast
