+++ presentation/src/main/java/com/going/presentation/todo/allocator/TodoAllocatorAdapter.kt
@@ -5,7 +5,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
 import com.going.domain.entity.response.TodoAllocatorModel
 import com.going.presentation.databinding.ItemTodoNameBinding
-import com.going.ui.extension.ItemDiffCallback
+import com.going.ui.util.ItemDiffCallback
 
 class TodoAllocatorAdapter(
     private val isCompleted: Boolean
