+++ presentation/src/main/java/com/going/presentation/todo/create/TodoCreateActivity.kt
@@ -16,7 +16,7 @@ import com.going.presentation.R
 import com.going.presentation.databinding.ActivityTodoCreateBinding
 import com.going.presentation.todo.TodoActivity.Companion.EXTRA_TRIP_ID
 import com.going.ui.base.BaseActivity
-import com.going.ui.extension.UiState
+import com.going.ui.state.UiState
 import com.going.ui.extension.setOnSingleClickListener
 import com.going.ui.extension.toast
 import dagger.hilt.android.AndroidEntryPoint
