+++ presentation/src/main/java/com/going/presentation/todo/create/TripParticipantAdapter.kt
@@ -5,7 +5,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
 import com.going.domain.entity.response.TripParticipantModel
 import com.going.presentation.databinding.ItemTodoCreateNameBinding
-import com.going.ui.extension.ItemDiffCallback
+import com.going.ui.util.ItemDiffCallback
 
 class TripParticipantAdapter(
     private val isFixed: Boolean,
