+++ presentation/src/main/java/com/going/presentation/todo/detail/PrivateDetailActivity.kt
@@ -8,7 +8,7 @@ import com.going.presentation.R
 import com.going.presentation.databinding.ActivityPrivateDetailBinding
 import com.going.presentation.todo.detail.PublicDetailActivity.Companion.EXTRA_TODO_ID
 import com.going.ui.base.BaseActivity
-import com.going.ui.extension.EnumUiState
+import com.going.ui.state.EnumUiState
 import com.going.ui.extension.setOnSingleClickListener
 import com.going.ui.extension.toast
 import dagger.hilt.android.AndroidEntryPoint
