+++ presentation/src/main/java/com/going/presentation/todo/detail/PrivateDetailViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.going.domain.repository.TodoRepository
-import com.going.ui.extension.EnumUiState
+import com.going.ui.state.EnumUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
