+++ presentation/src/main/java/com/going/presentation/todo/mytodo/MyTodoViewModel.kt
@@ -5,8 +5,8 @@ import androidx.lifecycle.viewModelScope
 import com.going.domain.entity.response.MyTripInfoModel
 import com.going.domain.entity.response.TodoModel
 import com.going.domain.repository.TodoRepository
-import com.going.ui.extension.EnumUiState
-import com.going.ui.extension.UiState
+import com.going.ui.state.EnumUiState
+import com.going.ui.state.UiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
