+++ presentation/src/main/java/com/going/presentation/todo/ourtodo/OurTodoViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.viewModelScope
 import com.going.domain.entity.response.OurTripInfoModel
 import com.going.domain.entity.response.TodoModel
 import com.going.domain.repository.TodoRepository
-import com.going.ui.extension.UiState
+import com.going.ui.state.UiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
